[test] Move test_core helper functions to new file
