+++ test/CMakeLists.txt
@@ -25,6 +25,7 @@ if(BUILD_TESTING)
     ###########
     add_executable(test_core
         test_core.cpp
+        test_core_helper.cpp
         test_request.cpp)
 
     target_link_libraries(test_core
